+++ Speed_Control.h
@@ -35,4 +35,4 @@
 #include "Uart_helper.h"
 #include "Read_Encoder.h"
 
-void Speed_Control(double Speed, double uSpeed);
+void Speed_Control(double Speed, int uSpeed);
